+++ src/components/features/detailPage/FestivalDetailPage.jsx
