+++ src/components/features/detailPage/atoms/InfoElement.jsx
